refactor: 동아리 카테고리 도메인을 하위 패키지로 분리
