+++ src/main/java/in/koreatech/koin/domain/club/model/Club.java
@@ -10,6 +10,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import in.koreatech.koin.common.model.BaseEntity;
+import in.koreatech.koin.domain.club.category.model.ClubCategory;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
