+++ src/main/java/in/koreatech/koin/domain/club/service/ClubService.java
@@ -46,7 +46,7 @@ import in.koreatech.koin.domain.club.exception.ClubLikeNotFoundException;
 import in.koreatech.koin.domain.club.exception.ClubManagerAlreadyException;
 import in.koreatech.koin.domain.club.model.Club;
 import in.koreatech.koin.domain.club.model.ClubBaseInfo;
-import in.koreatech.koin.domain.club.model.ClubCategory;
+import in.koreatech.koin.domain.club.category.model.ClubCategory;
 import in.koreatech.koin.domain.club.model.ClubEvent;
 import in.koreatech.koin.domain.club.model.ClubEventImage;
 import in.koreatech.koin.domain.club.model.ClubEventSubscription;
@@ -59,7 +59,7 @@ import in.koreatech.koin.domain.club.model.ClubRecruitmentSubscription;
 import in.koreatech.koin.domain.club.model.ClubSNS;
 import in.koreatech.koin.domain.club.model.redis.ClubCreateRedis;
 import in.koreatech.koin.domain.club.model.redis.ClubHotRedis;
-import in.koreatech.koin.domain.club.repository.ClubCategoryRepository;
+import in.koreatech.koin.domain.club.category.repository.ClubCategoryRepository;
 import in.koreatech.koin.domain.club.repository.ClubEventImageRepository;
 import in.koreatech.koin.domain.club.repository.ClubEventRepository;
 import in.koreatech.koin.domain.club.repository.ClubEventSubscriptionRepository;
